// src/shared/dataprovider/api/model/agentRemittance.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type AgentRemittance = 1 | 2 | 3;

export const AgentRemittance = {
    NUMBER_1: 1 as AgentRemittance,
    NUMBER_2: 2 as AgentRemittance,
    NUMBER_3: 3 as AgentRemittance
};