// src/shared/dataprovider/api/model/agentRemittanceListDataDTO.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AgentBankType } from './agentBankType';
import { AgentRemittance } from './agentRemittance';
import { ShortenedReceipt } from './shortenedReceipt';
import { ShortenedUploadedId } from './shortenedUploadedId';
import { ShortenedUploadedLetter } from './shortenedUploadedLetter';

export interface AgentRemittanceListDataDTO { 
    TransactionId?: number;
    Status?: AgentRemittance;
    StatusDescription?: string;
    TimeStamp?: Date;
    RemittanceId?: string;
    AgentId?: number;
    Agentname?: string;
    BankWithdrawalFee?: number;
    USDTWithdrawalFee?: number;
    CashOnPickupWithdrawalFee?: number;
    RequestedAmount?: number;
    TotalAmount?: number;
    TransactionFee?: number;
    WithdrawalCount?: number;
    DateRemitted?: Date;
    MOP?: number;
    MOPBankName?: string;
    SourceID?: number;
    SourceName?: string;
    Receipt?: Array<ShortenedReceipt>;
    UploadedId?: Array<ShortenedUploadedId>;
    AuhtorizationLetters?: Array<ShortenedUploadedLetter>;
    ReferenceNo?: string;
    Requestee?: string;
    ApprovedById?: number;
    ApprovedByName?: string;
    Remarks?: string;
    Method?: AgentBankType;
    MethodDescription?: string;
    RequestedByAdminId?: number;
    RequestedByAdmin?: string;
}